Move test/core/ -> test/test_core/ to make room for tests for lib/core/
